django/requirements.txt
Django
djangorestframework
django-cors-headers
mysqlclient
uwsgi
pyyaml
uritemplate
google-api-python-client
oauth2client
httplib2
opencv-python
pillow
# To use Auth Token
django-redis-cache
django-redis-sessions
# OpenCV(Edit movie)
opencv-python
Pillow
# .Env
django-environ
redis
